+++ R/set_blanket.R
@@ -56,7 +56,7 @@ set_blanket <- function(
     col_palette_d = jumble,
     col_palette_na_d = "#cdc5bfff",
     col_palette_c = viridisLite::mako(n = 9, direction = -1),
-    col_palette_na_c = "#cdc5bfff",
+    col_palette_na_c = "#988f88ff",
     theme = light_mode_r(orientation = "x"),
     ...
 ) {