+++ R/weave.R
@@ -3,12 +3,12 @@ ggblanket_global <- new.env(parent = emptyenv())
 
 ggblanket_global$mode <- NULL
 ggblanket_global$col_palette_d <- NULL
-ggblanket_global$col_palette_na_d <- NULL
+ggblanket_global$col_palette_d_na <- NULL
 ggblanket_global$col_palette_c <- NULL
-ggblanket_global$col_palette_na_c <- NULL
+ggblanket_global$col_palette_c_na <- NULL
 ggblanket_global$theme <- NULL
 ggblanket_global$col_palette_o <- NULL
-ggblanket_global$col_palette_na_o <- NULL
+ggblanket_global$col_palette_o_na <- NULL
 
 #' Set a default mode
 #'
@@ -168,8 +168,8 @@ weave_col_palette_d <- function(new = jumble, na = "#cdc5bfff") {
   ggblanket_global$col_palette_d <- new2
   invisible(old)
 
-  old <- ggblanket_global$col_palette_na_d
-  ggblanket_global$col_palette_na_d <- na
+  old <- ggblanket_global$col_palette_d_na
+  ggblanket_global$col_palette_d_na <- na
   invisible(old)
 
   if (rlang::is_null(new)) {
@@ -198,7 +198,7 @@ weave_col_palette_d <- function(new = jumble, na = "#cdc5bfff") {
 
 #' Set a default ordinal colour palette
 #'
-#' @param new Colour palette function to use for ordinal scale, e.g. `scales::pal_viridis()`).
+#' @param new A colour palette function to use for ordinal scale, e.g. `scales::pal_viridis()`).
 #' @param na A default colour for NA on a discrete scale. A hex code or name.
 #'
 #' @noRd
@@ -212,8 +212,8 @@ weave_col_palette_o <- function(new = scales::pal_viridis(option = "G", directio
   ggblanket_global$col_palette_o <- new
   invisible(old)
 
-  old <- ggblanket_global$col_palette_na_o
-  ggblanket_global$col_palette_na_o <- na
+  old <- ggblanket_global$col_palette_o_na
+  ggblanket_global$col_palette_o_na <- na
   invisible(old)
 }
 
@@ -238,8 +238,8 @@ weave_col_palette_c <- function(new = viridisLite::mako(n = 9, direction = -1),
   ggblanket_global$col_palette_c <- new
   invisible(old)
 
-  old <- ggblanket_global$col_palette_na_c
-  ggblanket_global$col_palette_na_c <- na
+  old <- ggblanket_global$col_palette_c_na
+  ggblanket_global$col_palette_c_na <- na
   invisible(old)
 
   options(
@@ -284,14 +284,14 @@ get_col_palette_c <- function() ggblanket_global$col_palette_c
 #' Get the default discrete NA colour
 #' @description Get the currently set default discrete NA colour.
 #' @noRd
-get_col_palette_na_d <- function() ggblanket_global$col_palette_na_d
+get_col_palette_d_na <- function() ggblanket_global$col_palette_d_na
 
 #' Get the default ordinal NA colour
 #' @description Get the currently set default ordinal NA colour.
 #' @noRd
-get_col_palette_na_o <- function() ggblanket_global$col_palette_na_o
+get_col_palette_o_na <- function() ggblanket_global$col_palette_o_na
 
 #' Get the default continuous NA colour
 #' @description Get the currently set default continuous NA colour.
 #' @noRd
-get_col_palette_na_c <- function() ggblanket_global$col_palette_na_c
+get_col_palette_c_na <- function() ggblanket_global$col_palette_c_na