+++ NEWS.md
@@ -13,6 +13,7 @@
 * Breaking: removed `replace_seq`.
 * Added `col_drop` and `facet_drop` arguments.
 * Fixed bug relating to `gg_bin_2d` scales.
+* Updated the default NA colour used for continuous colour scales to `#988f88ff"` (i.e. `colorspace::darken(grey, 0.25)`).
 
 # ggblanket 8.0.0