+++ NEWS.md
@@ -4,8 +4,8 @@
 * Breaking: renamed `*_title` to `*_label`.
 * Breaking: renamed `titles_to_case` to `label_to_case`.
 * Breaking: removed `weave_*` functions.
-* Fixed bug where `grey_mode_*`/`dark_mode_*` were missing legend functionality.
 * Breaking: removed `geom_linewidth` and `geom_size` from `weave_geom_aes` and `set_blanket`.
+* Fixed bug where `grey_mode_*`/`dark_mode_*` were missing legend functionality.
 * Improved scales by using `scales::breaks_extended()`.
 * In `dark_mode_*`, fixed incorrect `axis_line_colour` default.
 * In `*_mode_*`, updated caption colour default.